+++ R/MetaDataDiagnostics.R
@@ -1,4 +1,4 @@
-# Copyright 2021 Observational Health Data Sciences and Informatics
+# Copyright 2022 Observational Health Data Sciences and Informatics
 #
 # This file is part of CohortDiagnostics
 #
@@ -33,7 +33,7 @@
   }
   sql <- SqlRender::loadRenderTranslateSql(
     "OrphanCodes.sql",
-    packageName = "CohortDiagnostics",
+    packageName = utils::packageName(),
     dbms = connection@dbms,
     tempEmulationSchema = tempEmulationSchema,
     vocabulary_database_schema = vocabularyDatabaseSchema,
@@ -82,7 +82,7 @@
   sql <-
     SqlRender::loadRenderTranslateSql(
       "DropOrphanConceptTempTables.sql",
-      packageName = "CohortDiagnostics",
+      packageName = utils::packageName(),
       dbms = connection@dbms,
       tempEmulationSchema = tempEmulationSchema
     )
@@ -110,7 +110,7 @@ createConceptCountsTable <- function(connectionDetails = NULL,
   sql <-
     SqlRender::loadRenderTranslateSql(
       "CreateConceptCountTable.sql",
-      packageName = "CohortDiagnostics",
+      packageName = utils::packageName(),
       dbms = connection@dbms,
       tempEmulationSchema = tempEmulationSchema,
       cdm_database_schema = cdmDatabaseSchema,
@@ -119,4 +119,81 @@ createConceptCountsTable <- function(connectionDetails = NULL,
       table_is_temp = conceptCountsTableIsTemp
     )
   DatabaseConnector::executeSql(connection, sql)
+}
+
+saveDatabaseMetaData <- function(databaseId,
+                                 databaseName,
+                                 databaseDescription,
+                                 exportFolder,
+                                 vocabularyVersionCdm,
+                                 vocabularyVersion) {
+  ParallelLogger::logInfo("Saving database metadata")
+  startMetaData <- Sys.time()
+  database <- dplyr::tibble(
+    databaseId = databaseId,
+    databaseName = dplyr::coalesce(databaseName, databaseId),
+    description = dplyr::coalesce(databaseDescription, databaseId),
+    vocabularyVersionCdm = !!vocabularyVersionCdm,
+    vocabularyVersion = !!vocabularyVersion,
+    isMetaAnalysis = 0
+  )
+  writeToCsv(data = database,
+             fileName = file.path(exportFolder, "database.csv"))
+  delta <- Sys.time() - startMetaData
+  writeLines(paste(
+    "Saving database metadata took",
+    signif(delta, 3),
+    attr(delta, "units")
+  ))
+}
+
+getCdmVocabularyVersion <- function(connection, cdmDatabaseSchema) {
+  vocabularyVersionCdm <- NULL
+  tryCatch({
+    vocabularyVersionCdm <-
+      DatabaseConnector::renderTranslateQuerySql(
+        connection = connection,
+        sql = "select * from @cdm_database_schema.cdm_source;",
+        cdm_database_schema = cdmDatabaseSchema,
+        snakeCaseToCamelCase = TRUE
+      ) %>%
+      dplyr::tibble()
+  }, error = function(...) {
+    warning("Problem getting vocabulary version. cdm_source table not found in the database.")
+    if (connection@dbms == "postgresql") { #this is for test that automated testing purpose
+      DatabaseConnector::dbExecute(connection, "ABORT;")
+    }
+  })
+
+  if (all(!is.null(vocabularyVersionCdm),
+          nrow(vocabularyVersionCdm) > 0,
+          'vocabularyVersion' %in% colnames(vocabularyVersionCdm))) {
+    if (nrow(vocabularyVersionCdm) > 1) {
+      warning('Please check ETL convention for OMOP cdm_source table. It appears that there is more than one row while only one is expected.')
+    }
+    vocabularyVersionCdm <- vocabularyVersionCdm %>%
+      dplyr::rename(vocabularyVersionCdm = .data$vocabularyVersion) %>%
+      dplyr::pull(vocabularyVersionCdm) %>%
+      max() %>%
+      unique()
+  } else {
+    warning("Problem getting vocabulary version. cdm_source table either does not have data, or does not have the field vocabulary_version.")
+    vocabularyVersionCdm <- "Unknown"
+  }
+
+  return(vocabularyVersionCdm)
+}
+
+
+getVocabularyVersion <- function(connection, vocabularyDatabaseSchema) {
+  DatabaseConnector::renderTranslateQuerySql(
+    connection = connection,
+    sql = "select * from @vocabulary_database_schema.vocabulary where vocabulary_id = 'None';",
+    vocabulary_database_schema = vocabularyDatabaseSchema,
+    snakeCaseToCamelCase = TRUE
+  ) %>%
+    dplyr::tibble() %>%
+    dplyr::rename(vocabularyVersion = .data$vocabularyVersion) %>%
+    dplyr::pull(.data$vocabularyVersion) %>%
+    unique()
 }